# api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_florinel-chis_gophercrm_internal_models.Configuration:
    properties:
      category:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.ConfigurationCategory'
      created_at:
        type: string
      default_value:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      is_read_only:
        description: Read-only configs cannot be modified via API
        type: boolean
      is_system:
        description: System configs cannot be deleted
        type: boolean
      key:
        type: string
      type:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.ConfigurationType'
      updated_at:
        type: string
      valid_values:
        description: JSON array of valid values for validation
        type: string
      value:
        type: string
    type: object
  github_com_florinel-chis_gophercrm_internal_models.ConfigurationCategory:
    enum:
    - general
    - leads
    - customers
    - tickets
    - tasks
    - security
    - integration
    - ui
    type: string
    x-enum-varnames:
    - CategoryGeneral
    - CategoryLeads
    - CategoryCustomers
    - CategoryTickets
    - CategoryTasks
    - CategorySecurity
    - CategoryIntegration
    - CategoryUI
  github_com_florinel-chis_gophercrm_internal_models.ConfigurationType:
    enum:
    - string
    - boolean
    - integer
    - float
    - json
    - array
    type: string
    x-enum-varnames:
    - ConfigTypeString
    - ConfigTypeBoolean
    - ConfigTypeInteger
    - ConfigTypeFloat
    - ConfigTypeJSON
    - ConfigTypeArray
...
  github_com_florinel-chis_gophercrm_internal_utils.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  github_com_florinel-chis_gophercrm_internal_utils.APIMeta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_florinel-chis_gophercrm_internal_utils.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
      meta:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIMeta'
      success:
        type: boolean
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handler.SetConfigurationRequest:
    properties:
      value:
        items:
          type: integer
        type: array
    required:
    - value
    type: object
...
info:
  contact: {}
  description: CRM API for managing users, leads, customers, tickets, tasks and API
    keys.
  title: GopherCRM API
  version: "1.0"
paths:
...
  /configurations:
    get:
      description: Retrieve every configuration entry, ordered by category and key
        (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Configurations retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  properties:
                    configurations:
                      items:
                        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Configuration'
                      type: array
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "403":
          description: Forbidden - Admin role required
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List all configurations
      tags:
      - configurations
  /configurations/{key}:
    get:
      description: Retrieve a single configuration entry by its key (admin only)
      parameters:
      - description: Configuration key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Configuration retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Configuration'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "403":
          description: Forbidden - Admin role required
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "404":
          description: Configuration not found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get a configuration by key
      tags:
      - configurations
    put:
      consumes:
      - application/json
      description: Set the value of an existing configuration and return the updated
        entry (admin only). Read-only configurations are rejected, as are values outside
        the entry's valid_values constraint.
      parameters:
      - description: Configuration key
        in: path
        name: key
        required: true
        type: string
      - description: New configuration value. Any JSON value is accepted, falsy ones
          (false, 0, \
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SetConfigurationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Configuration updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Configuration'
              type: object
        "400":
          description: Missing or null value, read-only configuration, or a value
            outside the entry's valid_values constraint
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "403":
          description: Forbidden - Admin role required
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "404":
          description: Configuration not found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Update a configuration value
      tags:
      - configurations
  /configurations/{key}/reset:
    post:
      description: Restore a configuration's value from its stored default and return
        the updated entry (admin only). Read-only configurations are rejected.
      parameters:
      - description: Configuration key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Configuration reset successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Configuration'
              type: object
        "400":
          description: Configuration is read-only
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "403":
          description: Forbidden - Admin role required
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "404":
          description: Configuration not found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Reset a configuration to its default value
      tags:
      - configurations
  /configurations/category/{category}:
    get:
      description: Retrieve all configurations belonging to a category (admin only).
        An unrecognised category yields an empty list rather than an error.
      parameters:
      - description: Configuration category
        enum:
        - general
        - leads
        - customers
        - tickets
        - tasks
        - security
        - integration
        - ui
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Configurations retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  properties:
                    configurations:
                      items:
                        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Configuration'
                      type: array
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "403":
          description: Forbidden - Admin role required
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List configurations in a category
      tags:
      - configurations
  /configurations/ui:
    get:
      description: 'Retrieve the configurations the frontend is allowed to read: the
        whole ui category, general.company_name, and a synthetic leads.conversion.allowed_statuses
        entry. Available to any authenticated user.'
      produces:
      - application/json
      responses:
        "200":
          description: Configurations retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  properties:
                    configurations:
                      items:
                        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Configuration'
                      type: array
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List UI-safe configurations
      tags:
      - configurations
...
securityDefinitions:
  ApiKeyAuth:
    description: API key issued via POST /api-keys. The plaintext key is shown only
      once at creation.
    in: header
    name: ApiKey
    type: apiKey
  BearerAuth:
    description: 'JWT access token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"